third_party/WebKit/LayoutTests/editing/selection/android-longtap-not-select-empty.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="log"></div>
<p id="p1"><span id="span1">paragraph A</span></p>
<p>paragraph B</p>
<script>
test(function () {
    var boundsP1 = p1.getBoundingClientRect();
    var boundsSpan1 = span1.getBoundingClientRect();

    window.eventSender.gestureLongPress((boundsSpan1.right + boundsP1.right) / 2, (boundsSpan1.top + boundsSpan1.bottom) / 2);
    assert_equals(window.getSelection().toString(), "", "long tap on empty area between span and p");

});
</script>
